WIP: ALL BROKE - figuring out verification... again
