+++ lib/session.js
@@ -70,13 +70,6 @@ function parsePrivateKey(PRIVATE_KEY, keypair) {
   return keypair.public;
 }
 
-// TODO set as part of creation
-// { idClaims: { exp: '24h' }
-let defaultIdMaxAge = "24h";
-let defaultAccessMaxAge = "1h";
-let defaultRefreshMaxAge = "1d";
-let defaultTrustedMaxAge = "7d";
-
 /**
  * @param {string} issuer
  * @param {JwsPriv} PRIVATE_KEY
@@ -84,32 +77,19 @@ let defaultTrustedMaxAge = "7d";
  * @param {String} myOptions.cookiePath
  */
 LibAuth.create = function (issuer, PRIVATE_KEY, myOptions) {
+  // TODO set as part of creation
+  // { idClaims: { exp: '24h' }
+  let defaultIdMaxAge = "24h";
+  let defaultAccessMaxAge = "1h";
+  let defaultRefreshMaxAge = "1d";
+  let defaultTrustedMaxAge = "7d";
+
   let opts = {
     issuer: issuer,
-    oidc: {},
-    credentials: {
-      basic: true,
-      username: "username",
-      password: "password",
-    },
-    challenge: {
-      idByteCount: 4,
-      /** @type {import('crypto').BinaryToTextEncoding} */
-      idEncoding: "base64",
-      maxAge: "24h",
-      maxAttempts: 5,
-      receiptByteCount: 16,
-      /** @type {import('crypto').BinaryToTextEncoding} */
-      receiptEncoding: "base64",
-    },
-    oauth2: {},
-    challengeMaxAge: 0,
-    challengeMaxAttempts: 0,
     authnParam: "authn",
     cookiePath: myOptions.cookiePath,
     // assigned elsewhere
     secret: "",
-    loginUrl: "",
   };
 
   // Cookie Stuff
@@ -293,7 +273,15 @@ LibAuth.create = function (issuer, PRIVATE_KEY, myOptions) {
 
   /** @param {any} _opts // TODO */
   function _credentials(_opts) {
-    let credOpts = Object.assign(opts.credentials, _opts);
+    let credOpts = Object.assign(
+      {},
+      {
+        basic: true,
+        username: "username",
+        password: "password",
+      },
+      _opts || {},
+    );
 
     /** @type {import('express').Handler} */
     async function _credentialRoutes(req, res, next) {
@@ -527,25 +515,13 @@ LibAuth.create = function (issuer, PRIVATE_KEY, myOptions) {
 
   /** @param {any} _opts // TODO */
   function _oauth2(_opts) {
-    opts.oauth2 = Object.assign(opts.oauth2, _opts);
+    // TODO copy OIDC more or less
 
-    let _oauth2Routes = {
-      /** @type {import('express').Handler} */
-      "github.com": function (req, res, next) {},
-    };
-
-    // TODO pick one!
-    //@ts-ignore
-    var _gh = opts.oauth2["github.com"];
-    if (_gh?.clientSecret) {
-      let ghRoutes = require("./oauth2/github.com/").create({
-        _gh,
-        opts,
-      });
-      _oauth2Routes["github.com"] = ghRoutes;
-    }
-
-    return _oauth2Routes;
+    let ghRoutes = require("./oauth2/github.com/").create({
+      _gh: opts,
+      opts,
+    });
+    return ghRoutes;
   }
 
   function _wellKnown() {
@@ -568,8 +544,6 @@ LibAuth.create = function (issuer, PRIVATE_KEY, myOptions) {
 
     return wellKnown;
   }
-
-  async function _setClaims(req, res) {}
 
   /**
    * @param {MyAccessClaims | MyIdClaims} claims
