+++ lib/verifier.js
@@ -32,36 +32,38 @@ let parseDuration = require("./parse-duration.js");
 
 var C = module.exports;
 
-C.create = function ({
-  // important
-  iss = "",
-  secret = "",
-  /** @type {import('./memory-store.js').MemoryStore} */
-  //@ts-ignore
-  store,
+C.create = function (libauth, libOpts, _pluginOpts) {
+  let defaultOpts = {
+    // optional
+    coolDownMs: 250,
+    idByteCount: 4,
+    /** @type {import('crypto').BinaryToTextEncoding} */
+    idEncoding: "base64",
+    duration: "24h",
+    maxAge: "", // deprecated
+    // TODO rename to retries?
+    maxAttempts: 5,
+    receiptByteCount: 16,
+    /** @type {import('crypto').BinaryToTextEncoding} */
+    receiptEncoding: "base64",
+  };
 
-  // optional
-  coolDownMs = 250,
-  idByteCount = 4,
-  /** @type {import('crypto').BinaryToTextEncoding} */
-  idEncoding = "base64",
-  duration = "24h",
-  maxAge = "", // deprecated
-  // TODO rename to retries?
-  maxAttempts = 5,
-  receiptByteCount = 16,
-  /** @type {import('crypto').BinaryToTextEncoding} */
-  receiptEncoding = "base64",
-  authnParam = "authn", // TODO use same name as elsewhere
-}) {
-  // TODO document: 300s, 5m, 12h, 30d
-  if (!duration && maxAge) {
-    duration = maxAge;
+  let pluginOpts = Object.assign({}, defaultOpts, _pluginOpts || {});
+  if (!pluginOpts.store) {
+    console.warn(
+      "[libauth] Warn: no 'store' given, falling back to in-memory (single-system only) store",
+    );
+    pluginOpts.store = require("./memory-store.js");
   }
-  let durationMs = parseDuration(duration);
-  let HMAC_SECRET = secret;
 
-  let Challenge = {};
+  // TODO document: 300s, 5m, 12h, 30d
+  if (!pluginOpts.duration && pluginOpts.maxAge) {
+    pluginOpts.duration = pluginOpts.maxAge;
+  }
+  let durationMs = parseDuration(pluginOpts.duration);
+  let HMAC_SECRET = libOpts.secret;
+
+  let MagicLink = {};
 
   /**
    * @param {{
@@ -77,7 +79,7 @@ C.create = function ({
    * @param {any} opts
    * @returns {challenge}
    */
-  Challenge.create = function (identifier, req, opts = {}) {
+  MagicLink.create = function (identifier, req, opts = {}) {
     let ua = req?.userAgent || (req?.headers && req.headers["user-agent"]);
     let ip = req?.ip;
     let state = req?.state || req?.body?.state;
@@ -100,16 +102,19 @@ C.create = function ({
     // window is shrunk from 20 minutes to 10m or 5m
     //
     // https://therootcompany.com/blog/how-many-bits-of-entropy-per-character/
-    let { code, receipt, id } = Challenge._rndCode(4, "hex");
+    let { code, receipt, id } = MagicLink._rndCode(4, "hex");
 
     let d = new Date();
-    let _duration = opts?.duration || opts?.maxAge || duration;
+    let _duration = opts?.duration || opts?.maxAge || pluginOpts.duration;
     let expiration = new Date(d.valueOf() + parseDuration(_duration));
+
+    // TODO we should pass back one grand representation,
+    // pluck the receipt from it, and have the code separate
+
     /** @type challenge */
-    let c = {
+    let order = {
       id: id,
       receipt: receipt, //
-      code: code,
       state: state,
       identifier: {
         type: identifier.type,
@@ -122,19 +127,50 @@ C.create = function ({
       ordered_at: d.toISOString(),
       ordered_by: ua,
       ordered_ip: ip,
-      iss: opts?.iss || iss, //
+      iss: libOpts.issuer,
       duration: _duration,
     };
 
-    return c;
+    // for the notify handler
+    let authnParams = {
+      strategy: "challenge",
+      code: c.code,
+      order: {
+        id: c.id,
+        state: c.state,
+        identifier: {
+          type: c.identifier?.type || c.type, // email
+          value: c.identifier?.value || c.value, // john.doe@gmail.com
+        },
+        type: c.type, // email
+        value: c.value, // john.doe@gmail.com
+        userAgent: c.ordered_by,
+        //@ts-ignore
+        issuer: opts?.iss || libOpts.issuer,
+        //@ts-ignore
+        iss: opts?.iss || libOpts.issuer,
+        // for the client
+      },
+      receipt: {
+        id: c.id,
+        duration: c.duration,
+        receipt: c.receipt,
+        expires_at: c.expires_at,
+      },
+    };
+
+    return {
+      order: order,
+      code: code,
+    };
   };
 
   /**
    * Sanitizes and returns challenge object
    * @param {string} id
    */
-  Challenge.get = async function (id) {
-    let c = await store.get(id);
+  MagicLink.get = async function (id) {
+    let c = await pluginOpts.store.get(id);
     if (!c) {
       throw E.CODE_NOT_FOUND();
     }
@@ -151,7 +187,7 @@ C.create = function ({
    * @param {string} id
    * @param {any} c
    */
-  Challenge.set = async function (id, c) {
+  MagicLink.set = async function (id, c) {
     c = Object.assign({}, c);
     if ("undefined" !== typeof c.code) {
       c.code = undefined;
@@ -161,7 +197,7 @@ C.create = function ({
         c[k] = undefined;
       }
     });
-    await store.set(id, c);
+    await pluginOpts.store.set(id, c);
   };
 
   /** @type {Object.<string, boolean>} */
@@ -174,21 +210,21 @@ C.create = function ({
    * @param {import('express').Request} [req]
    * @returns {Promise<challenge>}
    */
-  Challenge.check = async function (id, code = "", req = undefined) {
+  MagicLink.check = async function (id, code = "", req = undefined) {
     if (!id) {
       throw E.DEVELOPER_ERROR("'id' is missing from the query parameters");
     }
 
-    let c = await Challenge.get(id);
+    let c = await MagicLink.get(id);
 
     if (code) {
-      await Challenge._check(id, c, code, {
+      await MagicLink._check(id, c, code, {
         noDoubleRedeem: true,
         noDoubleExchange: true,
       });
     } else {
       // we can still check after redemption
-      Challenge._assertIsFresh(c, {
+      MagicLink._assertIsFresh(c, {
         noDoubleRedeem: false,
         noDoubleExchange: true,
         failedValidation: false,
@@ -204,11 +240,11 @@ C.create = function ({
    * @param {string} code
    * @param {import('express').Request} req
    */
-  Challenge.redeem = async function (id, code, req) {
+  MagicLink.redeem = async function (id, code, req) {
     let finalize = req.body.finalize;
-    let c = await Challenge.get(id);
+    let c = await MagicLink.get(id);
 
-    await Challenge._check(id, c, code, {
+    await MagicLink._check(id, c, code, {
       noDoubleRedeem: true,
       noDoubleExchange: false,
     });
@@ -222,7 +258,7 @@ C.create = function ({
       c.exchanged_by = "";
       c.exchanged_ip = "";
     }
-    await store.set(id, c);
+    await pluginOpts.store.set(id, c);
 
     return c;
   };
@@ -233,14 +269,14 @@ C.create = function ({
    * @param {import('express').Request} req
    * @returns {Promise<challenge>}
    */
-  Challenge.exchange = async function (id, receipt, req) {
+  MagicLink.exchange = async function (id, receipt, req) {
     if (!id || !receipt) {
       throw E.DEVELOPER_ERROR(
         "'id' and/or 'receipt' is missing from the request body",
       );
     }
 
-    let c = await Challenge.get(id);
+    let c = await MagicLink.get(id);
 
     if (!c.verified_at) {
       // TODO better message and error code
@@ -248,7 +284,7 @@ C.create = function ({
         "a challenge code exchange was requested before the challenge code was submitted",
       );
     }
-    Challenge._assertIsFresh(c, {
+    MagicLink._assertIsFresh(c, {
       noDoubleRedeem: false,
       noDoubleExchange: true,
       failedValidation: false,
@@ -265,7 +301,7 @@ C.create = function ({
 
     c.exchanged_at = new Date().toISOString();
     // TODO audit log
-    await Challenge.set(id, c);
+    await MagicLink.set(id, c);
 
     return c;
   };
@@ -275,10 +311,10 @@ C.create = function ({
    * @param {import('express').Request} req
    * @returns {Promise<challenge>}
    */
-  Challenge.cancel = async function (id, req) {
-    let c = await Challenge.get(id);
+  MagicLink.cancel = async function (id, req) {
+    let c = await MagicLink.get(id);
     // the code isn't valid, but it's not invalid either - we don't need it
-    Challenge._assertIsFresh(c, {
+    MagicLink._assertIsFresh(c, {
       noDoubleRedeem: true,
       noDoubleExchange: true,
       failedValidation: false,
@@ -288,30 +324,10 @@ C.create = function ({
     c.canceled_by = req.headers["user-agent"];
     c.canceled_ip = req.ip;
 
-    await store.set(id, c);
-    //await Challenge.set(id, c);
+    await pluginOpts.store.set(id, c);
+    //await MagicLink.set(id, c);
 
     return c;
-  };
-
-  /**
-   * @param {Object} opts
-   * @param {String} opts.iss
-   * @param {String} opts.secret
-   * @param {String} opts.authnParam
-   */
-  // for internal use by libauth
-  // (so that issuer doesn't have to be set manually)
-  Challenge.setDefaults = function (opts) {
-    if (!iss && opts.iss) {
-      iss = opts.iss;
-    }
-    if (!secret && opts.secret) {
-      secret = opts.secret;
-    }
-    if (!authnParam && opts.authnParam) {
-      authnParam = opts.authnParam;
-    }
   };
 
   /**
@@ -321,7 +337,7 @@ C.create = function ({
    * @param {string} code
    * @param {AssertOpts} assertOpts
    */
-  Challenge._check = async function (id, c, code, assertOpts) {
+  MagicLink._check = async function (id, c, code, assertOpts) {
     // An attacker could grant himself hundreds or thousands of extra attempts
     // by firing off many requests in parallel - the database might read
     // `attempts = 0` 1000 times and then write `attempts = 1` 1000 times, and
@@ -330,12 +346,12 @@ C.create = function ({
     // To prevent this disallow parallel requests.
     // (note: a scalable server system will need a more sophisticated approach)
     if (attempts[id]) {
-      await C._sleep(coolDownMs);
+      await C._sleep(pluginOpts.coolDownMs);
       throw E.ENHANCE_YOUR_CALM();
     }
 
     attempts[id] = true;
-    let err = await Challenge._incrementOnFailure(c, code, assertOpts).catch(
+    let err = await MagicLink._incrementOnFailure(c, code, assertOpts).catch(
       Object,
     );
 
@@ -354,7 +370,7 @@ C.create = function ({
    * @param {String} code
    * @param {AssertOpts} assertOpts
    */
-  Challenge._incrementOnFailure = async function (c, code, assertOpts) {
+  MagicLink._incrementOnFailure = async function (c, code, assertOpts) {
     if (!code) {
       throw E.DEVELOPER_ERROR("'code' is missing from the request parameters");
     }
@@ -363,11 +379,11 @@ C.create = function ({
       HMAC_SECRET,
       c.receipt || "",
       code,
-      receiptByteCount,
-      receiptEncoding,
+      pluginOpts.receiptByteCount,
+      pluginOpts.receiptEncoding,
     );
     assertOpts.failedValidation = true === pass;
-    Challenge._assertIsFresh(c, assertOpts);
+    MagicLink._assertIsFresh(c, assertOpts);
     if (pass) {
       return;
     }
@@ -378,7 +394,7 @@ C.create = function ({
       c.attempts = 0;
     }
     c.attempts += 1;
-    await store.set(c.id, c);
+    await pluginOpts.store.set(c.id, c);
     throw E.CODE_RETRY();
   };
 
@@ -389,28 +405,33 @@ C.create = function ({
    * @throws {Error}
    */
   // TODO c should maintain durationMs, etc
-  Challenge._assertIsFresh = function (c, assertOpts) {
-    return C.__assertIsFresh(c, durationMs, maxAttempts, assertOpts);
+  MagicLink._assertIsFresh = function (c, assertOpts) {
+    return C.__assertIsFresh(c, durationMs, pluginOpts.maxAttempts, assertOpts);
   };
 
   /**
    * @param {number} bytes
    * @param {BufferEncoding} enc
    */
-  Challenge._rndCode = function (bytes, enc) {
+  MagicLink._rndCode = function (bytes, enc) {
     let code = rnd(bytes, enc);
     let receipt = C._hashify(
       HMAC_SECRET,
       code,
-      receiptByteCount,
-      receiptEncoding,
+      pluginOpts.receiptByteCount,
+      pluginOpts.receiptEncoding,
     );
-    let id = C._hashify(HMAC_SECRET, receipt, idByteCount, idEncoding);
+    let id = C._hashify(
+      HMAC_SECRET,
+      receipt,
+      pluginOpts.idByteCount,
+      pluginOpts.idEncoding,
+    );
 
     return { code, receipt, id };
   };
 
-  return Challenge;
+  return MagicLink;
 };
 
 C._sleep = async function sleep(n = 0) {
